refactor(models/product): change price typr for integer

+++ src/models/Cosmetic.java
@@ -5,7 +5,7 @@ public class Cosmetic extends Product {
     private String type;
     private Boolean containsSunProtectionFactor;
 
-    public Cosmetic(String name, Integer price, String brand, String type, Boolean containsSunProtectionFactor) {
+    public Cosmetic(String name, float price, String brand, String type, Boolean containsSunProtectionFactor) {
         super(name, price);
         this.brand = brand;
         this.type = type;
